Updates to the chain for generation of autoplaylists. Improvements to the LINQ queries to (hopefully) reduce it all down to a single query Note: This still doesn't compile.

+++ DolomiteModel/PublicRepresentations/AutoPlaylistRule.cs
@@ -6,7 +6,7 @@ namespace DolomiteModel.PublicRepresentations
     public class AutoPlaylistRule
     {
         [DataMember]
-        public int? Id { get; set; }
+        public long? Id { get; set; }
 
         /// <summary>
         /// The metadata field the rule is based on
